Add implementacion del Kardex en doCompra, doVenta Fix bugs

+++ AlphaInventory_Proyecto/test/Temp/SUBTEST INVENTARIO/model/Usuario.h
@@ -19,7 +19,7 @@ private:
     string nombre;
     string apellidos;
     string tipoDocumento;
-    int numDocumento;
+    long int numDocumento;
     string contrasena;
     string tipoUsuario;
